created a new branch for refactoring the repo, in this I created a new folder for translated graphs and for tests. I even renamed the main.py to application.py
